return ContextText on sendTweet failure
A common failure in Apps Script gives "Send tweet failure. Error was: {"name":"Exception"}", which is generated on line 344 but doesn't correctly pass on the error message.

The answer here: https://stackoverflow.com/a/66253113 does what I'm proposing in a sideband test to get the real error message. It can easily be returned instead of null on line 345.